+++ app/src/main/java/com/concordium/wallet/ui/cis2/SendTokenViewModel.kt
@@ -25,7 +25,6 @@ import com.concordium.wallet.data.model.AccountBalance
 import com.concordium.wallet.data.model.AccountData
 import com.concordium.wallet.data.model.AccountNonce
 import com.concordium.wallet.data.model.GlobalParamsWrapper
-import com.concordium.wallet.data.model.InputEncryptedAmount
 import com.concordium.wallet.data.model.SubmissionData
 import com.concordium.wallet.data.model.Token
 import com.concordium.wallet.data.model.TransactionOutcome
@@ -40,6 +39,7 @@ import com.concordium.wallet.ui.account.common.accountupdater.AccountUpdater
 import com.concordium.wallet.ui.common.BackendErrorHandler
 import com.concordium.wallet.util.DateTimeUtil
 import com.concordium.wallet.util.Log
+import com.concordium.wallet.util.TokenUtil
 import com.concordium.wallet.util.toBigInteger
 import com.concordium.wallet.util.toHex
 import kotlinx.coroutines.CoroutineScope
@@ -300,23 +300,7 @@ class SendTokenViewModel(application: Application) : AndroidViewModel(applicatio
         val accountRepository =
             AccountRepository(WalletDatabase.getDatabase(getApplication()).accountDao())
         val account = accountRepository.findByAddress(accountAddress)
-        val atDisposal =
-            account?.getAtDisposalWithoutStakedOrScheduled(account.totalUnshieldedBalance)
-                ?: BigInteger.ZERO
-        return Token(
-            "",
-            "CCD",
-            "",
-            null,
-            false,
-            "",
-            "",
-            true,
-            (account?.totalBalance ?: BigInteger.ZERO),
-            atDisposal,
-            "",
-            "CCD"
-        )
+        return TokenUtil.getCCDToken(account)
     }
 
     fun getCipherForBiometrics(): Cipher? {